# docs/元数据服务/openapi.json
{
  "components": {
    "schemas": {
      "_table__guid_Response": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object",
            "properties": {
              "owner": {
                "format": "string",
                "description": "负责人",
                "type": "string"
              },
              "typeName": {
                "format": "string",
                "description": "类型名称",
                "type": "string"
              },
              "tableRows": {
                "format": "string",
                "description": "表行数",
                "type": "string"
              },
              "dataLength": {
                "format": "string",
                "description": "数据长度",
                "type": "string"
              },
              "indexLength": {
                "format": "string",
                "description": "索引长度",
                "type": "string"
              },
              "comment": {
                "format": "string",
                "description": "备注信息",
                "type": "string"
              },
              "qualifiedName": {
                "format": "string",
                "description": "限定名",
                "type": "string"
              },
              "description": {
                "format": "string",
                "description": "描述信息",
                "type": "string"
              },
              "labels": {
                "format": "string",
                "description": "标签",
                "type": "string"
              },
              "classification": {
                "format": "string",
                "description": "分类",
                "type": "string"
              },
              "createTime": {
                "format": "string",
                "description": "创建时间",
                "type": "string"
              },
              "db": {
                "description": "db信息(object)",
                "type": "object",
                "properties": {
                  "guid": {
                    "format": "string",
                    "description": "No comments found.",
                    "type": "string"
                  },
                  "typeName": {
                    "format": "string",
                    "description": "No comments found.",
                    "type": "string"
                  },
                  "displayText": {
                    "format": "string",
                    "description": "No comments found.",
                    "type": "string"
                  }
                }
              },
              "columns": {
                "description": "参考实体",
                "type": "array",
                "items": {
                  "properties": {
                    "any object": {
                      "description": "any object.(object)",
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "__guids_Response": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object"
          }
        }
      },
      "_listRequest": {
        "properties": {
          "query": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "typeName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "classification": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "limit": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer"
          },
          "offse": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer"
          }
        }
      },
...
      "_detail__typeName___guid_Response": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object",
            "properties": {
              "owner": {
                "format": "string",
                "description": "负责人",
                "type": "string"
              },
              "typeName": {
                "format": "string",
                "description": "类型名称",
                "type": "string"
              },
              "tableRows": {
                "format": "string",
                "description": "表行数",
                "type": "string"
              },
              "dataLength": {
                "format": "string",
                "description": "数据长度",
                "type": "string"
              },
              "indexLength": {
                "format": "string",
                "description": "索引长度",
                "type": "string"
              },
              "comment": {
                "format": "string",
                "description": "备注信息",
                "type": "string"
              },
              "qualifiedName": {
                "format": "string",
                "description": "限定名",
                "type": "string"
              },
              "description": {
                "format": "string",
                "description": "描述信息",
                "type": "string"
              },
              "dataType": {
                "format": "string",
                "description": "数据类型",
                "type": "string"
              },
              "isPrimaryKey": {
                "format": "string",
                "description": "是否是主键",
                "type": "string"
              },
              "defaultValue": {
                "format": "string",
                "description": "默认值",
                "type": "string"
              },
              "labels": {
                "format": "string",
                "description": "标签",
                "type": "string"
              },
              "classification": {
                "format": "string",
                "description": "分类",
                "type": "string"
              },
              "createTime": {
                "format": "string",
                "description": "创建时间",
                "type": "string"
              },
              "db": {
                "description": "db信息(object)",
                "type": "object",
                "properties": {
                  "guid": {
                    "format": "string",
                    "description": "No comments found.",
                    "type": "string"
                  },
                  "typeName": {
                    "format": "string",
                    "description": "No comments found.",
                    "type": "string"
                  },
                  "displayText": {
                    "format": "string",
                    "description": "No comments found.",
                    "type": "string"
                  }
                }
              },
              "relationshipAttributes": {
                "description": "实体信息(map data)",
                "type": "object"
              }
            }
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://172.20.0.13:8782/mtd"
    }
  ],
  "openapi": "3.0.3",
  "paths": {
...
    "/detail/{typeName}/{guid}": {
      "get": {
        "summary": "根据guid和typeName获取数据库的元数据信息",
        "deprecated": false,
        "description": "根据guid和typeName获取数据库的元数据信息",
        "tags": [
          "MtdAtlasController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_detail__typeName___guid_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"1ircio\",\n  \"tips\": \"p81wls\",\n  \"data\": {\n    \"owner\": \"ifgrve\",\n    \"typeName\": \"rochel.emard\",\n    \"tableRows\": \"uk9tje\",\n    \"dataLength\": \"nwkdkb\",\n    \"indexLength\": \"jh9mhh\",\n    \"comment\": \"emkie4\",\n    \"qualifiedName\": \"rochel.emard\",\n    \"description\": \"5byr8v\",\n    \"dataType\": \"34z54r\",\n    \"isPrimaryKey\": \"ybmoge\",\n    \"defaultValue\": \"nefuqf\",\n    \"labels\": \"yryg7s\",\n    \"classification\": \"r5djdc\",\n    \"createTime\": \"2021-10-25 16:17:38\",\n    \"db\": {\n      \"guid\": \"59\",\n      \"typeName\": \"rochel.emard\",\n      \"displayText\": \"hmzhle\"\n    },\n    \"relationshipAttributes\": {\n      \"mapKey\": {}\n    }\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "getDetailByGuid",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "实体类型",
            "required": true,
            "example": "rochel.emard",
            "name": "typeName"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "实体id",
            "required": true,
            "example": "59",
            "name": "guid"
          }
        ]
      }
    },
    "/audit/{guid}": {
      "get": {
        "summary": "根据guid获取数据库的技术数据操作信息",
        "deprecated": false,
        "description": "根据guid获取数据库的技术数据操作信息",
        "tags": [
          "MtdAtlasController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_audit__guid_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"i15fa0\",\n  \"tips\": \"es7t9p\",\n  \"data\": [\n    {\n      \"entityId\": \"59\",\n      \"timestamp\": 1635149857944,\n      \"user\": \"864b4v\",\n      \"action\": \"ENTITY_CREATE\",\n      \"details\": \"2noeli\",\n      \"eventKey\": \"h62vab\",\n      \"entity\": {\n        \"typeName\": \"rochel.emard\",\n        \"attributes\": {\n          \"mapKey\": {}\n        },\n        \"guid\": \"59\",\n        \"homeId\": \"59\",\n        \"isProxy\": true,\n        \"isIncomplete\": true,\n        \"provenanceType\": 176,\n        \"status\": \"ACTIVE\",\n        \"createdBy\": \"i0t9dr\",\n        \"updatedBy\": \"3awptz\",\n        \"createTime\": \"2021-10-25 16:17:38\",\n        \"updateTime\": \"2021-10-25 16:17:38\",\n        \"version\": 864,\n        \"relationshipAttributes\": {\n          \"mapKey\": {}\n        },\n        \"classifications\": [\n          {\n            \"typeName\": \"rochel.emard\",\n            \"attributes\": {\n              \"mapKey\": {}\n            },\n            \"entityGuid\": \"59\",\n            \"entityStatus\": \"ACTIVE\",\n            \"propagate\": true,\n            \"validityPeriods\": [\n              {\n                \"startTime\": \"2021-10-25 16:17:37\",\n                \"endTime\": \"2021-10-25 16:17:37\",\n                \"timeZone\": \"drx2em\"\n              }\n            ],\n            \"removePropagationsOnEntityDelete\": true\n          }\n        ],\n        \"meanings\": [\n          {\n            \"termGuid\": \"59\",\n            \"relationGuid\": \"59\",\n            \"description\": \"c834zq\",\n            \"displayText\": \"bqu3am\",\n            \"expression\": \"l3yfy0\",\n            \"createdBy\": \"cfspzn\",\n            \"steward\": \"z6omfr\",\n            \"source\": \"dkwmci\",\n            \"confidence\": 695,\n            \"status\": \"DISCOVERED\"\n          }\n        ],\n        \"customAttributes\": {\n          \"mapKey\": \"0zv28g\"\n        },\n        \"businessAttributes\": {\n          \"mapKey\": {}\n        },\n        \"labels\": [\n          \"psh3af\"\n        ],\n        \"pendingTasks\": [\n          \"h57vmu\"\n        ]\n      },\n      \"type\": \"ENTITY_AUDIT_V1\"\n    }\n  ]\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "getAuditByGuid",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "实体id",
            "required": true,
            "example": "59",
            "name": "guid"
          }
        ]
      }
    },
    "/{guids}": {
      "delete": {
        "summary": "根据guid删除元数据,多个guid 使用英文逗号分割",
        "deprecated": false,
        "description": "根据guid删除元数据,多个guid 使用英文逗号分割",
        "tags": [
          "MtdAtlasController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/__guids_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"hpwpoy\",\n  \"tips\": \"jelx4d\",\n  \"data\": {\n    \"waring\": \"You may have used non-display generics.\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "deleteEntitiesByGuids",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "No comments found.",
            "required": true,
            "example": "kvukrf",
            "name": "guids"
          }
        ]
      }
    },
...
    "/table/{guid}": {
      "get": {
        "summary": "根据guid获取表的元数据详情",
        "deprecated": false,
        "description": "根据guid获取表的元数据详情",
        "tags": [
          "MtdAtlasController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_table__guid_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"1tt744\",\n  \"tips\": \"hbip8z\",\n  \"data\": {\n    \"owner\": \"o6u98e\",\n    \"typeName\": \"rochel.emard\",\n    \"tableRows\": \"f8p31u\",\n    \"dataLength\": \"c5ev60\",\n    \"indexLength\": \"cmgatq\",\n    \"comment\": \"uhez9w\",\n    \"qualifiedName\": \"rochel.emard\",\n    \"description\": \"obu4ea\",\n    \"labels\": \"vvt6g2\",\n    \"classification\": \"9ccguk\",\n    \"createTime\": \"2021-10-25 16:17:38\",\n    \"db\": {\n      \"guid\": \"59\",\n      \"typeName\": \"rochel.emard\",\n      \"displayText\": \"nyfb05\"\n    },\n    \"columns\": [\n      {}\n    ]\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "getTableDetailByGuid",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "元数据唯一id",
            "required": true,
            "example": "59",
            "name": "guid"
          }
        ]
      }
    },
    "/db/{guid}": {
      "get": {
        "summary": "根据guid获取数据库的元数据信息",
        "deprecated": false,
        "description": "根据guid获取数据库的元数据信息",
        "tags": [
          "MtdAtlasController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_db__guid_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"nvdb9l\",\n  \"tips\": \"iwrs23\",\n  \"data\": {\n    \"owner\": \"j5e7fl\",\n    \"typeName\": \"rochel.emard\",\n    \"comment\": \"0orai0\",\n    \"qualifiedName\": \"rochel.emard\",\n    \"description\": \"afoav5\",\n    \"dataType\": \"vqm8x1\",\n    \"labels\": \"cbtvzl\",\n    \"classification\": \"m3g5hg\",\n    \"createTime\": \"2021-10-25 16:17:38\",\n    \"tables\": [\n      {}\n    ]\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "getDbDetailByGuid",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "元数据唯一id",
            "required": true,
            "example": "59",
            "name": "guid"
          }
        ]
      }
    },
...
    "/detail/qname/{qualifiedName}/{typename}": {
      "get": {
        "summary": "根据guid获取元数据技术数据详情",
        "deprecated": false,
        "description": "根据guid获取元数据技术数据详情",
        "tags": [
          "MtdAtlasController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_detail_qname__qualifiedName___typename_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"ltoy56\",\n  \"tips\": \"oiyd8c\",\n  \"data\": {\n    \"typeName\": \"rochel.emard\",\n    \"attributes\": {\n      \"mapKey\": {}\n    },\n    \"guid\": \"59\",\n    \"homeId\": \"59\",\n    \"isProxy\": true,\n    \"isIncomplete\": true,\n    \"provenanceType\": 206,\n    \"status\": \"ACTIVE\",\n    \"createdBy\": \"dbn23c\",\n    \"updatedBy\": \"6am5a5\",\n    \"createTime\": \"2021-10-25 16:17:38\",\n    \"updateTime\": \"2021-10-25 16:17:38\",\n    \"version\": 32,\n    \"relationshipAttributes\": {\n      \"mapKey\": {}\n    },\n    \"classifications\": [\n      {\n        \"typeName\": \"rochel.emard\",\n        \"attributes\": {\n          \"mapKey\": {}\n        },\n        \"entityGuid\": \"59\",\n        \"entityStatus\": \"ACTIVE\",\n        \"propagate\": true,\n        \"validityPeriods\": [\n          {\n            \"startTime\": \"2021-10-25 16:17:37\",\n            \"endTime\": \"2021-10-25 16:17:37\",\n            \"timeZone\": \"2s9rz4\"\n          }\n        ],\n        \"removePropagationsOnEntityDelete\": true\n      }\n    ],\n    \"meanings\": [\n      {\n        \"termGuid\": \"59\",\n        \"relationGuid\": \"59\",\n        \"description\": \"680dby\",\n        \"displayText\": \"bwmzdl\",\n        \"expression\": \"veaul4\",\n        \"createdBy\": \"zbsii1\",\n        \"steward\": \"b8h26e\",\n        \"source\": \"93w36a\",\n        \"confidence\": 632,\n        \"status\": \"DISCOVERED\"\n      }\n    ],\n    \"customAttributes\": {\n      \"mapKey\": \"akzw74\"\n    },\n    \"businessAttributes\": {\n      \"mapKey\": {}\n    },\n    \"labels\": [\n      \"1qbpd7\"\n    ],\n    \"pendingTasks\": [\n      \"0u8mrc\"\n    ]\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "getDetailByQName",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "元数据",
            "required": true,
            "example": "rochel.emard",
            "name": "qualifiedName"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "类型名称",
            "required": true,
            "example": "rochel.emard",
            "name": "typename"
          }
        ]
      }
    },
...
    "/column/{guid}": {
      "get": {
        "summary": "根据guid获取列元数据详情",
        "deprecated": false,
        "description": "根据guid获取列元数据详情",
        "tags": [
          "MtdAtlasController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_column__guid_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"be22r5\",\n  \"tips\": \"h2vwep\",\n  \"data\": {\n    \"owner\": \"q0f16t\",\n    \"typeName\": \"rochel.emard\",\n    \"displayName\": \"rochel.emard\",\n    \"comment\": \"o6uirb\",\n    \"qualifiedName\": \"rochel.emard\",\n    \"description\": \"7jsslj\",\n    \"dataType\": \"zrnrbo\",\n    \"isPrimaryKey\": \"jvf2t2\",\n    \"defaultValue\": \"9f74k7\",\n    \"createTime\": \"2021-10-25 16:17:38\",\n    \"labels\": \"ds8ai0\",\n    \"classification\": \"4k3v67\",\n    \"table\": {\n      \"guid\": \"59\",\n      \"typeName\": \"rochel.emard\",\n      \"displayText\": \"axn0ky\"\n    }\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "getColumnDetailByGuid",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "元数据唯一id",
            "required": true,
            "example": "59",
            "name": "guid"
          }
        ]
      }
    },
...
    "/list": {
      "post": {
        "summary": "查询元数据列表",
        "deprecated": false,
        "description": "查询元数据列表",
        "tags": [
          "MtdAtlasController"
        ],
        "requestBody": {
          "content": {
            "application/json; charset\u003dutf-8": {
              "schema": {
                "$ref": "#/components/schemas/_listRequest"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"query\": \"82fmjg\",\n  \"typeName\": \"rochel.emard\",\n  \"classification\": \"uz43c0\",\n  \"limit\": 10,\n  \"offse\": 932\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_listResponse"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"yvutno\",\n  \"tips\": \"ojg97y\",\n  \"data\": [\n    {\n      \"guid\": \"59\",\n      \"typeName\": \"rochel.emard\",\n      \"displayName\": \"rochel.emard\",\n      \"qualifiedName\": \"rochel.emard\",\n      \"createTime\": \"2021-10-25 16:17:38\"\n    }\n  ]\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "searchList",
        "parameters": []
      }
    },
...
    "/detail/{guid}": {
      "get": {
        "summary": "根据guid获取数据库的技术元数据",
        "deprecated": false,
        "description": "根据guid获取数据库的技术元数据",
        "tags": [
          "MtdAtlasController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_detail__guid_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"hflhjr\",\n  \"tips\": \"qsqf83\",\n  \"data\": {\n    \"owner\": \"ph97h5\",\n    \"typeName\": \"rochel.emard\",\n    \"tableRows\": \"v5a7hq\",\n    \"dataLength\": \"o9kdiu\",\n    \"indexLength\": \"113vkh\",\n    \"comment\": \"94j5b7\",\n    \"qualifiedName\": \"rochel.emard\",\n    \"description\": \"xfs8a4\",\n    \"dataType\": \"2bfiq4\",\n    \"isPrimaryKey\": \"4j1av0\",\n    \"defaultValue\": \"vw4bw0\",\n    \"labels\": \"n3k25j\",\n    \"classification\": \"thq6sh\",\n    \"createTime\": \"2021-10-25 16:17:38\",\n    \"db\": {\n      \"guid\": \"59\",\n      \"typeName\": \"rochel.emard\",\n      \"displayText\": \"luxpy2\"\n    },\n    \"relationshipAttributes\": {\n      \"mapKey\": {}\n    }\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "getDetailByGuid",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "元数据id",
            "required": true,
            "example": "59",
            "name": "guid"
          }
        ]
      }
    },
...
    "/criteria/list": {
      "post": {
        "summary": "多条件查询元数据列表",
        "deprecated": false,
        "description": "多条件查询元数据列表",
        "tags": [
          "MtdAtlasController"
        ],
        "requestBody": {
          "content": {
            "application/json; charset\u003dutf-8": {
              "schema": {
                "$ref": "#/components/schemas/_criteria_listRequest"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"query\": \"4jyqox\",\n  \"typeName\": \"rochel.emard\",\n  \"classification\": \"t379k7\",\n  \"limit\": 10,\n  \"offset\": 1,\n  \"typeNameValue\": [\n    \"4cn9qk\"\n  ],\n  \"nameValue\": [\n    \"dzn0tg\"\n  ],\n  \"labelsValue\": [\n    \"p845gs\"\n  ]\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_criteria_listResponse"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"klbgq1\",\n  \"tips\": \"g26a18\",\n  \"data\": [\n    {\n      \"guid\": \"59\",\n      \"typeName\": \"rochel.emard\",\n      \"displayName\": \"rochel.emard\",\n      \"qualifiedName\": \"rochel.emard\",\n      \"createTime\": \"2021-10-25 16:17:38\"\n    }\n  ]\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "searchListByCriteria",
        "parameters": []
      }
    },
...
  },
  "info": {
    "title": "dm-metadata",
    "version": "1.0.0"
  }
}